refactor: split server.ts into focused modules under server/
